Lots of improvements!
Got a basic window setup, and got SFML to properly run withing the
window.  Right now it only renders a green circle, but it opens it up
for game graphics to be written to the screen.  Also integrated
keypresses to modify the controller area in ram, so the game can see
when keys are pressed.  Right now the keys are hardcoded in as WASD for
arrows, KM for AB, and VB as select/start.  Really excited!